ucombo file and some typing
